old-tests/webapps/WebStorage/tests/approved/storage_session_getitem.html

<!DOCTYPE HTML>
<html>
 <head>
  <title>Web Storage</title>
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
 </head>
 <body>
	<p>
		<h3>storage_session_getItem</h3>
	</p>
	
	<hr>

	<div id="log"></div>

	<script>

		if (('sessionStorage' in window) && window.sessionStorage !== null){
			sessionStorage.clear();
			sessionStorage.setItem("name", "user1");
			sessionStorage.setItem("age", "20");

			try {
				test(function() {
					assert_equals(sessionStorage.getItem("name"), "user1", "sessionStorage.getItem('name')")
				}, "getItem method test")
			} catch (e) {
				test(function()
				{
					assert_unreached("getItem method is failed.");
				}, "getItem method test")
			}
		} else {
			test(function()
			{
				assert_unreached("sessionStorage is not supported.");
			}, "sessionStorage support test")
		}	
	</script>

 </body>
</html>
